Redirect was added to save and delete buttons

+++ src/components/noteEditor/index.js
@@ -2,6 +2,7 @@ import { useContext, useState, useEffect } from "react";
 import GlobalContext from "../../GlobalContext";
 import { useParams, useHistory } from "react-router-dom";
 import { Prompt } from "react-router";
+import { Redirect } from "react-router-dom";
 import ReactQuill from "react-quill";
 import "react-quill/dist/quill.snow.css";
 import "./CustomQuillStyles.css";
@@ -12,6 +13,7 @@ import { colors } from "../../commonStyles/variables";
 
 const NoteEditor = () => {
   const [isChanged, setIsChanged] = useState(false);
+  const [redirect, setRedirect] = useState(false);
   const { notes, setNotes } = useContext(GlobalContext);
   const [currentNoteEditorState, setCurrentNoteEditorState] = useState(notes);
   const { id } = useParams();
@@ -49,45 +51,46 @@ const NoteEditor = () => {
       "Are you sure you would like to delete your note?"
     );
     if (confirmDelete) {
-      setIsChanged(false);
-      setTimeout(() => history.push("/"), 0);
-      setTimeout(() => setNotes(notes.filter((note) => note.id !== id)), 0);
+      setNotes(notes.filter((note) => note.id !== id));
+      setRedirect(true);
     } else {
       return;
     }
   };
 
   const handleSave = () => {
-    setIsChanged(false);
     setNotes(currentNoteEditorState);
-
-    setTimeout(() => history.push("/"), 0);
+    setRedirect(true);
   };
 
-  return (
-    <div className="custom-quill">
-      <Prompt
-        when={isChanged}
-        message="You have unsaved changes, are you sure you want to leave?"
-      />
-      <ReactQuill
-        theme="snow"
-        value={getNote().text}
-        onChange={(e) => handleChange(e)}
-      />
-      <div className="custom-quill-footer">
-        <div
-          className="trash-icon-wrapper"
-          onClick={(e) => handleDelete(e, id)}
-        >
-          <SvgIcon color={colors.red} size={"40px"}>
-            <IoMdTrash />
-          </SvgIcon>
+  if (redirect) {
+    return <Redirect to="/" />;
+  } else {
+    return (
+      <div className="custom-quill">
+        <Prompt
+          when={isChanged}
+          message="You have unsaved changes, are you sure you want to leave?"
+        />
+        <ReactQuill
+          theme="snow"
+          value={getNote().text}
+          onChange={(e) => handleChange(e)}
+        />
+        <div className="custom-quill-footer">
+          <div
+            className="trash-icon-wrapper"
+            onClick={(e) => handleDelete(e, id)}
+          >
+            <SvgIcon color={colors.red} size={"40px"}>
+              <IoMdTrash />
+            </SvgIcon>
+          </div>
+          <Button onClick={handleSave}>Save</Button>
         </div>
-        <Button onClick={handleSave}>Save</Button>
       </div>
-    </div>
-  );
+    );
+  }
 };
 
 export default NoteEditor;
